PLAT-283: fixing implemenataion of market metrics and adding a test
The catalog market metrics method was changed when we fixed performance in other catalog to dataframe methods such as catalog markets. It was supposed to be simplified, but ended up duplicating rows. The problem is described in more [detail here](https://coinmetrics.slack.com/archives/C02FG4L7GJ0/p1685650082512609). This MR:
* Reverts the catalog_market_metrics implementation back to the working version
* Adds a test to show it works as expected - it is creating one row for each frequencies present