DATA-34 Unnest Catalog V2 DataFrames
This MR lets catalog v2 dataframes be unnested and simplifies the CatalogV2 dataframe code.

**Background in addition to JIRA ticket details**

For Catalog V1, our approach to returning catalogs involved creating a separate data type for each catalog. This resulted in several instances of redundant code and made it expanding upon each new catalog data type laborious. For Catalog V2, we'll generalize all catalog returns as a single data type and account for the differences in structure across endpoint in how the CatalogV2DataCollection object is passed.

**Potential breaking changes**

Code that manually explode catalog v2 dataframes, specifically relying on referencing specific columns such as `frequencies` and `metrics` columns, will break. They are being replaced by their singular terms to represent the individual rows that are being unnested from those fields. Having said that, catalog v2 dataframes were not all that useful in their current state and were in an experimental stage, so I think it's fine to introduce this change.

**Tests**

To test the functionality of this code, each catalog v2 endpoint was called as a dataframe and list then matched against the expected fields.